+++ Gravitas.CollisionCoordination/Models/CollisionData.cs
@@ -4,7 +4,7 @@ namespace Gravitas.CollisionCoordination.Models
 {
     public class CollisionData
     {
-        public long? TicketId { get; set; } 
+        public int? TicketId { get; set; } 
         public List<string> PhoneList { get; set; }
         public List<string> EmailList { get; set; }
         public int? TemplateId { get; set; }
